[cleanup] reorganize rbac-related methods
